+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - A page control can be hidden when the value of the key parameter in the URL does not matches one of the list of `show_for_key` values in the list definition.
+- A new setting for `groupDefaultExpanded` can be used to set the number of initially expanded groups in a grid.
 ### Changed
 - Use library-specific Error class instead of RuntimeError.
 ### Fixed